Only update agency tables if changes are detected

+++ dataactvalidator/scripts/load_agencies.py
@@ -1,6 +1,8 @@
 import os
 import logging
 import requests
+import re
+import numpy as np
 
 import pandas as pd
 
@@ -10,8 +12,37 @@ from dataactcore.logging import configure_logging
 from dataactcore.models.domainModels import CGAC, SubTierAgency, FREC
 from dataactvalidator.health_check import create_app
 from dataactvalidator.scripts.loader_utils import clean_data
+from dataactbroker.helpers.pandas_helper import check_dataframe_diff
 
 logger = logging.getLogger(__name__)
+
+
+def extract_abbreviation(row):
+    """ Helper function for comparing datasets, this extracts the abbreviation from the db name
+
+        Args:
+            row: row in the dataframe representing the db model
+
+        Returns:
+            abbreviation extracted from agency_name or np.nan
+    """
+    abbr = re.findall('^.* \((.*)\)$', row['agency_name'])
+    if abbr and abbr[0] != 'nan':
+        return abbr[0]
+    else:
+        return np.nan
+
+
+def extract_name(row):
+    """ Helper function for comparing datasets, this extracts the name from the db name
+
+        Args:
+            row: row in the dataframe representing the db model
+
+        Returns:
+            name extracted from agency_name or np.nan
+    """
+    return row['agency_name'][:row['agency_name'].rindex('(')-1]
 
 
 def delete_missing_cgacs(models, new_data):
@@ -47,11 +78,12 @@ def update_cgacs(models, new_data):
             setattr(models[cgac_code], field, value)
 
 
-def load_cgac(file_name):
+def load_cgac(file_name, force_reload=False):
     """ Load CGAC (high-level agency names) lookup table.
 
         Args:
             file_name: path/url to the file to be read
+            force_reload: whether to reload regardless
     """
     sess = GlobalDB.db().session
     models = {cgac.cgac_code: cgac for cgac in sess.query(CGAC)}
@@ -69,12 +101,20 @@ def load_cgac(file_name):
     # de-dupe
     data.drop_duplicates(subset=['cgac_code'], inplace=True)
 
-    delete_missing_cgacs(models, data)
-    update_cgacs(models, data)
-    sess.add_all(models.values())
-    sess.commit()
+    # compare to existing content in table
+    diff_found  = check_dataframe_diff(data, CGAC, 'cgac_id', ['cgac_code'],
+                                      lambda_funcs=[('agency_abbreviation', extract_abbreviation),
+                                                    ('agency_name', extract_name)])
 
-    logger.info('%s CGAC records inserted', len(models))
+    if force_reload or diff_found:
+        delete_missing_cgacs(models, data)
+        update_cgacs(models, data)
+        sess.add_all(models.values())
+        sess.commit()
+
+        logger.info('%s CGAC records inserted', len(models))
+    else:
+        logger.info('No differences found, skipping cgac table reload.')
 
 
 def delete_missing_frecs(models, new_data):
@@ -115,11 +155,12 @@ def update_frecs(models, new_data, cgac_dict):
             setattr(models[frec_code], field, value)
 
 
-def load_frec(file_name):
+def load_frec(file_name, force_reload=False):
     """ Load FREC (high-level agency names) lookup table.
 
         Args:
             file_name: path/url to the file to be read
+            force_reload: whether to reload regardless
     """
     sess = GlobalDB.db().session
     models = {frec.frec_code: frec for frec in sess.query(FREC)}
@@ -137,18 +178,32 @@ def load_frec(file_name):
     )
     # de-dupe
     data = data[data.frec_cgac == 'TRUE']
+    data.drop(['frec_cgac'], axis=1, inplace=True)
     data.drop_duplicates(subset=['frec_code'], inplace=True)
-    # create foreign key dicts
+
     cgac_dict = {str(cgac.cgac_code): cgac.cgac_id for
                  cgac in sess.query(CGAC).filter(CGAC.cgac_code.in_(data["cgac_code"])).all()}
+    cgac_dict_flipped = {cgac_id: cgac_code for cgac_code, cgac_id in cgac_dict.items()}
 
-    # insert to db
-    delete_missing_frecs(models, data)
-    update_frecs(models, data, cgac_dict)
-    sess.add_all(models.values())
-    sess.commit()
+    # compare to existing content in table
+    extract_cgac = lambda row: cgac_dict_flipped[row['cgac_id']] if row['cgac_id'] in cgac_dict_flipped else np.nan
+    diff_found  = check_dataframe_diff(data, FREC, 'frec_id', ['frec_code'], fk_cols=['cgac_id'],
+                                       lambda_funcs=[('agency_abbreviation', extract_abbreviation),
+                                                     ('agency_name', extract_name),
+                                                     ('cgac_code', extract_cgac)])
+    if force_reload or diff_found:
+        # create foreign key dicts
 
-    logger.info('%s FREC records inserted', len(models))
+        # insert to db
+        delete_missing_frecs(models, data)
+        update_frecs(models, data, cgac_dict)
+        sess.add_all(models.values())
+        sess.commit()
+
+        logger.info('%s FREC records inserted', len(models))
+    else:
+        logger.info('No differences found, skipping frec table reload.')
+
 
 
 def delete_missing_sub_tier_agencies(models, new_data):
@@ -189,11 +244,12 @@ def update_sub_tier_agencies(models, new_data, cgac_dict, frec_dict):
             setattr(models[sub_tier_agency_code], field, value)
 
 
-def load_sub_tier_agencies(file_name):
+def load_sub_tier_agencies(file_name, force_reload=False):
     """ Load Sub Tier Agency (sub_tier-level agency names) lookup table.
 
         Args:
             file_name: path/url to the file to be read
+            force_reload: whether to reload regardless
     """
     sess = GlobalDB.db().session
     models = {sub_tier_agency.sub_tier_agency_code: sub_tier_agency for
@@ -221,22 +277,36 @@ def load_sub_tier_agencies(file_name):
     # create foreign key dicts
     cgac_dict = {str(cgac.cgac_code): cgac.cgac_id for
                  cgac in sess.query(CGAC).filter(CGAC.cgac_code.in_(data["cgac_code"])).all()}
+    cgac_dict_flipped = {cgac_id: cgac_code for cgac_code, cgac_id in cgac_dict.items()}
     frec_dict = {str(frec.frec_code): frec.frec_id for
                  frec in sess.query(FREC).filter(FREC.frec_code.in_(data["frec_code"])).all()}
+    frec_dict_flipped = {frec_id: frec_code for frec_code, frec_id in frec_dict.items()}
 
-    delete_missing_sub_tier_agencies(models, data)
-    update_sub_tier_agencies(models, data, cgac_dict, frec_dict)
-    sess.add_all(models.values())
-    sess.commit()
+    # compare to existing content in table
+    int_to_float = lambda row: float(row['priority'])
+    extract_cgac = lambda row: cgac_dict_flipped[row['cgac_id']] if row['cgac_id'] in cgac_dict_flipped else np.nan
+    extract_frec = lambda row: frec_dict_flipped[row['frec_id']] if row['frec_id'] in frec_dict_flipped else np.nan
+    diff_found  = check_dataframe_diff(data, SubTierAgency, 'sub_tier_agency_id', ['sub_tier_agency_code'],
+                                       fk_cols=['cgac_id', 'frec_id'],
+                                       lambda_funcs=[('priority', int_to_float), ('cgac_code', extract_cgac),
+                                                     ('frec_code', extract_frec)])
+    if force_reload or diff_found:
+        delete_missing_sub_tier_agencies(models, data)
+        update_sub_tier_agencies(models, data, cgac_dict, frec_dict)
+        sess.add_all(models.values())
+        sess.commit()
 
-    logger.info('%s Sub Tier Agency records inserted', len(models))
+        logger.info('%s Sub Tier Agency records inserted', len(models))
+    else:
+        logger.info('No differences found, skipping subtier table reload.')
 
 
-def load_agency_data(base_path):
+def load_agency_data(base_path, force_reload=False):
     """ Load agency data into the database
 
         Args:
             base_path: directory that contains the agency files
+            force_reload: whether to reload regardless
     """
     agency_codes_url = 'https://files.usaspending.gov/reference_data/agency_codes.csv'
     logger.info('Loading agency codes file from {}'.format(agency_codes_url))
@@ -249,11 +319,11 @@ def load_agency_data(base_path):
     # Parse data
     with create_app().app_context():
         logger.info('Loading CGAC')
-        load_cgac(agency_codes_file)
+        load_cgac(agency_codes_file, force_reload=force_reload)
         logger.info('Loading FREC')
-        load_frec(agency_codes_file)
+        load_frec(agency_codes_file, force_reload=force_reload)
         logger.info('Loading Sub Tier Agencies')
-        load_sub_tier_agencies(agency_codes_file)
+        load_sub_tier_agencies(agency_codes_file, force_reload=force_reload)
 
     # Delete file once we're done
     os.remove(agency_codes_file)
